soc: qcom: Add High-Speed Phy control to EUD driver
The Embedded USB Debugger (EUD) is a HS-USB on-chip hub to support the
debug and trace capabilities on Qualcomm devices. It sits in between
the usb port and the usb controller. Being a HS USB hub, it relies on
HS Phy for its functionality. Add HS phy support in the eud driver
and control the phy during eud enable/disable operations.

Change-Id: If807cf5f3e70b0ed9c169a8c0bb43ec7118e49d6
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>